Extract rotation status bar package
